requirements.txt
-e .
attrs==19.3.0
aws-cdk.aws-elasticsearch==1.36.1
aws-cdk.aws-events==1.36.1
aws-cdk.aws-iam==1.36.1
aws-cdk.aws-kinesisfirehose==1.36.1
aws-cdk.aws-kms==1.36.1
aws-cdk.aws-s3==1.36.1
aws-cdk.aws-ec2==1.36.1
aws-cdk.cloud-assembly-schema==1.36.1
aws-cdk.core==1.36.1
aws-cdk.cx-api==1.36.1
aws-cdk.region-info==1.36.1
cattrs==1.0.0
constructs==3.0.2
jsii==1.4.1
python-dateutil==2.8.1
six==1.14.0
typing-extensions==3.7.4.2
